[CLEAN] - Refactoring code
[*] Move ordinal to number into parsing modules

[*] Cleaning code
