+++ tests/utils.rs
@@ -1,5 +1,5 @@
 use anitomy_rust::utils::{
-    is_digit, is_hexa, normalize, ordinals_to_nb, remove_ignored_string, split_by_delimiter,
+    is_digit, is_hexa, normalize, remove_ignored_string, split_by_delimiter,
 };
 
 #[test]
@@ -75,12 +75,4 @@ fn split_regex() {
         "test",
     ];
     assert_eq!(e, split_by_delimiter(s, d))
-}
-
-#[test]
-fn ordinal() {
-    assert_eq!(ordinals_to_nb("First"), "1");
-    assert_eq!(ordinals_to_nb("second"), "2");
-    assert_eq!(ordinals_to_nb("5th"), "5");
-    assert_eq!(ordinals_to_nb("5the"), "");
 }
